chat-backend/requirements.txt
fastapi
uvicorn
python-socketio[asgi]
python-jose
passlib[argon2]
sqlalchemy
asyncpg
redis
httpx
pydantic
greenlet